pyproject.toml
[tool.poetry]
name = "database-api"
version = "0.1.0"
description = ""
authors = ["hawa1222 <hw97.business@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
aiomysql = "^0.2.0"
mysql-connector-python = "8.2.0"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
pandas = "^2.2.2"
cryptography = "^42.0.5"
pytest = "^8.1.1"
httpx = "^0.27.0"
python-jose = "3.3.0"
passlib = "1.7.4"
python-multipart = "0.0.9"
greenlet = "3.0.3"
bcrypt = "4.0.1"
python-dotenv = "^1.0.1"
slowapi = "^0.1.9"
requests = "^2.31.0"
pytest-asyncio = "^0.23.6"

[[tool.poetry.packages]]
include = "app"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"